@@ -7,9 +7,9 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "bootstrap": "^5.1.0",
-    "react": "^17.0.2",
+    "react": "^19.2.0",
     "react-bootstrap": "^2.0.0-beta.6",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.2.0",
     "react-icons": "^4.2.0",
     "react-scripts": "^5.0.1",
     "styled-components": "^5.3.0",
